Process event location in EventManager
